% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric.parse.R
\name{as.numeric.parse}
\alias{as.numeric.parse}
\title{Wrapper for the as.numeric function that prints entries that
fail to parse}
\usage{
\method{as.numeric}{parse}(x, ...)
}
\arguments{
\item{x}{string or vector to coerce to numeric}
}
\description{
Wrapper for the as.numeric function that prints entries that
fail to parse
}
\examples{
z <- as.numeric.parse(c(1:5, "String1",6:10,"String2"))
z
}
\keyword{as.numeric}
