% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demography}
\alias{demography}
\title{Demography data

Simulated demography for patients.}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{Subject}{Patient ID}
\item{GENDER_CODE}{Patient gender}
}
}
\usage{
demography
}
\description{
Demography data

Simulated demography for patients.
}
\keyword{datasets}
