% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug2_admin}
\alias{drug2_admin}
\title{Study agent 2 data

Simulated study agent for patients.}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{Subject}{Patient ID}
\item{TX2_DATE_INT}{Study agent 2 start date of patient on study}
}
}
\usage{
drug2_admin
}
\description{
Study agent 2 data

Simulated study agent for patients.
}
\keyword{datasets}
