% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment}
\alias{enrollment}
\title{Enrollment data

Simulated enrollment for patients.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{Subject}{Patient ID}
\item{COHORT}{Study cohort for patient}
\item{ENROL_DATE_INT}{Enrollment date of patient to study}
}
}
\usage{
enrollment
}
\description{
Enrollment data

Simulated enrollment for patients.
}
\keyword{datasets}
