% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ineligibility}
\alias{ineligibility}
\title{Simulated ineligibility for patients.}
\format{
A data frame with 11 rows and 2 variables:
\describe{
\item{Subject}{Patient ID}
\item{INELIGIBILITY_STATUS}{Recorded ineligibility status of patient to study}
}
}
\usage{
ineligibility
}
\description{
Simulated ineligibility for patients.
}
\keyword{datasets}
