% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_timeline_plot.R
\name{ae_timeline_plot}
\alias{ae_timeline_plot}
\title{Outputs AE timeline plots including just system organ class (AE category), or
system organ class and lower level term (AE detail)}
\usage{
ae_timeline_plot(
  subjID,
  subjID_ineligText = NULL,
  baseline_datasets,
  ae_dataset,
  ae_attribVars,
  ae_attribVarsName = NULL,
  ae_attribVarText = NULL,
  enrolDtVar,
  ae_detailVar,
  ae_categoryVar,
  ae_severityVar,
  ae_onsetDtVar,
  time_unit = c("day", "week", "month", "year"),
  include_ae_detail = T,
  ...
)
}
\arguments{
\item{subjID}{key identifier field for participant ID in data sets}

\item{subjID_ineligText}{character text that denotes participant IDs to exclude,
for example, c("New Subject") (if provided)}

\item{baseline_datasets}{list of data frames that contain baseline participant characteristics,
for example, list(enrollment_DF,demography_DF,ineligibility_DF)}

\item{ae_dataset}{data frame that contains subject AEs}

\item{ae_attribVars}{field(s) that denotes attribution to intervention under study,
for example, c("CTC_AE_ATTR_SCALE","CTC_AE_ATTR_SCALE_1") (if provided)}

\item{ae_attribVarsName}{character text that denotes name of interventions under study,
for example, c("Drug 1", "Drug 2") (if provided)}

\item{ae_attribVarText}{character text that denotes related attribution, for example
c("Definite", "Probable", "Possible") (if provided)}

\item{enrolDtVar}{field that denotes participant enrolment date (i.e. 10MAY2021)}

\item{ae_detailVar}{field that denotes participant AE detail (lower-level term)}

\item{ae_categoryVar}{field that denotes participant AE category (system organ class)}

\item{ae_severityVar}{field that denotes participant AE severity grade (numeric)}

\item{ae_onsetDtVar}{field that denotes participant AE onset date}

\item{time_unit}{character text that denotes time unit for desired timeline,
for example, could be one of c("day","week","month","year") (if provided)}

\item{include_ae_detail}{boolean that denotes if AE detail should be included
in timeline plot. Default is True}
}
\value{
ggplot object of AE timeline plot
}
\description{
Outputs AE timeline plots including just system organ class (AE category), or
system organ class and lower level term (AE detail)
}
\examples{
data("enrollment", "ae");
plotOut <- ae_timeline_plot(subjID="Subject",subjID_ineligText=c("New Subject","Test"),
  baseline_datasets=list(enrollment),
  ae_dataset=ae,
  ae_attribVars=c("CTC_AE_ATTR_SCALE","CTC_AE_ATTR_SCALE_1",
               "CTC_AE_ATTR_SCALE","CTC_AE_ATTR_SCALE_1",
               "CTC_AE_ATTR_SCALE"),
  ae_attribVarsName=c("Drug 1","Drug 2", "Drug 3","Drug 4", "Drug 5"),
  ae_attribVarText=c("Definite", "Probable", "Possible"),
  enrolDtVar="ENROL_DATE_INT",ae_detailVar="ae_detail",
  ae_categoryVar="ae_category",ae_severityVar="AE_SEV_GD",
  ae_onsetDtVar="AE_ONSET_DT_INT",time_unit="week")
  ggsave(paste("ae_detail_timeline_plot_example_2", ".png", sep=""), plotOut, width=6.4, height=18, device="png", scale = 1);
}
\keyword{plot}
