% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug1_admin}
\alias{drug1_admin}
\title{Study agent 1 data

Simulated study agent for patients.}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{Subject}{Patient ID}
\item{TX1_DATE_INT}{Study agent 1 start date of patient on study}
}
}
\usage{
drug1_admin
}
\description{
Study agent 1 data

Simulated study agent for patients.
}
\keyword{datasets}
