% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb_ccru.R
\name{dsmb_ccru}
\alias{dsmb_ccru}
\title{Outputs the three DSMB-CCRU AE summary tables in Excel format per UHN template}
\usage{
dsmb_ccru(
  protocol,
  setwd,
  title,
  comp = NULL,
  pi,
  presDate,
  cutDate,
  boundDate = NULL,
  subjID,
  enrollment_DF,
  demography_DF,
  ineligibility_DF,
  ae_DF,
  genderVar,
  ineligVar,
  enrolDtVar,
  ae_detailVar,
  ae_categoryVar,
  ae_severityVar,
  ae_onsetDtVar,
  ae_detailOtherText = NULL,
  ae_detailOtherVar = NULL,
  ae_verbatimVar = NULL,
  numSubj = NULL,
  ...
)
}
\arguments{
\item{protocol}{study protocol name (uppercase, no spaces permitted)}

\item{setwd}{working directory to write summary files to}

\item{title}{full character vector with name of study}

\item{comp}{comparison group (if provided)}

\item{pi}{character vector name of study principal investigator}

\item{presDate}{presentation date (i.e. 17NOV2023) for DSMB}

\item{cutDate}{recent cutoff date for AEs (i.e. 31AUG2023)}

\item{boundDate}{lower bound cutoff date for AEs (if provided)}

\item{subjID}{key identifier field for subject ID in data sets}

\item{enrollment_DF}{data set that contains enrollment variables}

\item{demography_DF}{data set that contains demography variables}

\item{ineligibility_DF}{data set that contains ineligibility variables}

\item{ae_DF}{data set that contains AEs}

\item{genderVar}{field that denotes participant gender}

\item{ineligVar}{field that denotes participant ineligibility (Yes, No)}

\item{enrolDtVar}{field that denotes participant enrolment date (i.e. 10MAY2021)}

\item{ae_detailVar}{field that denotes participant AE detail (lower-level term)}

\item{ae_categoryVar}{field that denotes participant AE category (system organ class)}

\item{ae_severityVar}{field that denotes participant AE severity grade (numeric)}

\item{ae_onsetDtVar}{field that denotes participant AE onset date}

\item{ae_detailOtherText}{character text that denotes referencing vebatim AE field,
for example, "Other, specify" (if provided)}

\item{ae_detailOtherVar}{field that denotes participant AE detail other (if provided)}

\item{ae_verbatimVar}{field that denotes participant AE detail verbatim (if provided)}

\item{numSubj}{vector to override value for number of subjects in summary (if provided)}
}
\value{
three Excel files containing DSMB-CCRU AE summary tables
}
\description{
Outputs the three DSMB-CCRU AE summary tables in Excel format per UHN template
}
\examples{
dsmb_ccru(protocol="CLINICAL_TRIAL",setwd="T:/My location/study",
   title="A feasibility study to evaluate a cancer drug",
   comp="COHORT",pi="Dr. PI",presDate="17NOV2023",cutDate="31AUG2023",boundDate=NULL,
   subjID="Subject",enrollment_DF=enrollment_DF,demography_DF=demography_DF,
   ineligibility_DF=ineligibility_DF,ae_DF=ae_DF,
   genderVar="GENDER_CODE",ineligVar="INELIGIBILITY_STATUS",enrolDtVar="ENROL_DATE_INT",
   ae_detailVar="ae_detail",ae_categoryVar="ae_category",
   ae_severityVar="AE_SEV_GD_STD",ae_onsetDtVar="AE_ONSET_DT_INT",
   ae_verbatimOtherText="Other, specify",ae_detailOtherVar="CTCAE5_LLT_NM",
   ae_verbatimVar="AE_VERBATIM_TRM_TXT",numSubj=c(4,5,6,7))
}
\keyword{dataframe}
