% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_mcmcglmm.R
\name{nice_mcmcglmm}
\alias{nice_mcmcglmm}
\title{Nice model output from MCMCglmm::MCMCglmm()}
\usage{
nice_mcmcglmm(mcmcglmm_object, dataset)
}
\arguments{
\item{mcmcglmm_object}{returned output from MCMCglmm()}

\item{data}{dataframe containing data}
}
\description{
Nice model output from MCMCglmm::MCMCglmm()
}
\examples{
mcmcglmm_mva <- nice_mcmcglmm(model1, Milk);
options(knitr.kable.NA = '');
knitr::kable(mcmcglmm_mva);
}
\keyword{dataframe}
